refactor: consolidate constants and remove duplicate Header barrel
