+++ src/app/api/cms/assets/[...slug]/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 
-import { directusApiToken, directusUrl } from '@/constant/env';
+import { directusApiToken, directusUrl } from '@/constants/env';
 
 const FILE_ID_RE =
   /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
