--- src/components/layout/Header.tsx
@@ -1,2 +1,0 @@
-export { Header } from './header/Header';
-export type { NavItemProps, SubItem } from './header/types';
