+++ src/components/layout/index.ts
@@ -1,5 +1,6 @@
 export * from './Container';
 export * from './Footer';
-export * from './Header';
 export * from './InfoBar';
 export * from './Layout';
+export { Header } from './header/Header';
+export type { NavItemProps, SubItem } from './header/types';
