+++ src/constants/__tests__/responsive-image-sizes.test.ts
@@ -1,7 +1,7 @@
 import {
   ImageBannerRole,
   responsiveImageSizes,
-} from '@/constant/responsive-image-sizes';
+} from '@/constants/responsive-image-sizes';
 
 describe('responsiveImageSizes', () => {
   it('should have the correct values for each role', () => {
